Commit de alteracoes

+++ src/br/com/project/locacao/calculo/Classificacao.java
@@ -1,5 +1,5 @@
 package br.com.project.locacao.calculo;
 
 public interface Classificacao {
-    int getClassificacao();
+    double getClassificacao();
 }
